+++ core/src/main/java/com/thefirstlineofcode/chalk/core/ITaskService.java
@@ -4,8 +4,8 @@ import com.thefirstlineofcode.basalt.xmpp.core.stanza.Stanza;
 
 public interface ITaskService {
 	void execute(ITask<?> task);
-	void setDefaultTimeout(int timeout);
-	int getDefaultTimeout();
+	void setDefaultTimeout(long timeout);
+	long getDefaultTimeout();
 	void setDefaultTimeoutHandler(ITimeoutHandler timeoutHandler);
 	<K extends Stanza, V> V execute(ISyncTask<K, V> task) throws ErrorException;
 }
