+++ core/src/main/java/com/thefirstlineofcode/chalk/core/IUnidirectionalStream.java
@@ -4,5 +4,5 @@ import com.thefirstlineofcode.basalt.xmpp.core.stanza.Stanza;
 
 public interface IUnidirectionalStream<T extends Stanza> {
 	void send(T stanza);
-	void send(T stanza, int timeout);
+	void send(T stanza, long timeout);
 }
